Move match_nodes into utils
